# grizzly-api-core-service/src/test/resources/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Build Time: 23/05/2019, 12:58",
    "version": "0.0.1-SNAPSHOT",
    "title": "Resource Manager Api"
  },
  "host": "localhost:8090",
  "basePath": "/",
...
  "paths": {
    "/allUsers": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer {genereated_token}",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/user/check/{email}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "existsByEmail",
        "operationId": "existsByEmailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer {genereated_token}",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "path",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  }
}